requirements.txt
# Audio Analysis
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0

# Metadata handling
mutagen>=1.46.0
tinytag>=1.9.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# File watching
watchdog>=3.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
click>=8.1.0
rich>=13.0.0
pandas>=2.0.0
schedule>=1.2.0

# Audio fingerprinting (optional)
pyacoustid>=1.2.0

# Machine Learning for genre/mood classification
scikit-learn>=1.3.0
joblib>=1.3.0
